+++ Ametrin/src/main/java/com/ametrinstudios/ametrin/data/provider/loot_table/ExtendedEntityLootSubProvider.java
@@ -1,7 +1,7 @@
 package com.ametrinstudios.ametrin.data.provider.loot_table;
 
-import net.minecraft.core.HolderLookup;
 import net.minecraft.data.loot.EntityLootSubProvider;
+import net.minecraft.data.loot.LootTableSubProvider;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.flag.FeatureFlags;
 import org.jetbrains.annotations.NotNull;
@@ -9,8 +9,8 @@ import org.jetbrains.annotations.NotNull;
 import java.util.stream.Stream;
 
 public abstract class ExtendedEntityLootSubProvider extends EntityLootSubProvider {
-    protected ExtendedEntityLootSubProvider(HolderLookup.Provider registries) {
-        super(FeatureFlags.REGISTRY.allFlags(), registries);
+    protected ExtendedEntityLootSubProvider(LootTableSubProvider.Context output) {
+        super(FeatureFlags.REGISTRY.allFlags(), output);
     }
 
     @Override