+++ Ametrin/src/main/java/com/ametrinstudios/ametrin/data/provider/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.ametrinstudios.ametrin.data.provider;
+
+import org.jspecify.annotations.NullMarked;