+++ Ametrin/src/main/java/com/ametrinstudios/ametrin/mixin/MixinBlockBehaviorProperties.java
@@ -13,6 +13,7 @@ import net.minecraft.world.level.block.state.properties.NoteBlockInstrument;
 import net.minecraft.world.level.material.MapColor;
 import net.minecraft.world.level.material.PushReaction;
 import net.minecraft.world.level.storage.loot.LootTable;
+import net.minecraft.world.phys.AABB;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
@@ -34,11 +35,14 @@ public abstract class MixinBlockBehaviorProperties implements IMixinBlockBehavio
     @Shadow float friction;
     @Shadow float speedFactor;
     @Shadow float jumpFactor;
+    @Shadow float bounceRestitution;
     @Shadow DependantName<Block, Optional<ResourceKey<LootTable>>> drops;
+    @Shadow DependantName<Block, String> descriptionId;
     @Shadow boolean canOcclude;
     @Shadow boolean isAir;
     @Shadow boolean ignitedByLava;
     @Shadow boolean liquid;
+    @Shadow boolean forceSolidOff;
     @Shadow boolean forceSolidOn;
     @Shadow PushReaction pushReaction;
     @Shadow boolean spawnTerrainParticles;
@@ -47,13 +51,15 @@ public abstract class MixinBlockBehaviorProperties implements IMixinBlockBehavio
     @Shadow BlockBehaviour.StateArgumentPredicate<EntityType<?>> isValidSpawn;
     @Shadow BlockBehaviour.StatePredicate isRedstoneConductor;
     @Shadow BlockBehaviour.StatePredicate isSuffocating;
-    @Shadow BlockBehaviour.StatePredicate isViewBlocking;
+    @Shadow BlockBehaviour.StateArgumentPredicate<AABB> isViewBlocking;
     @Shadow BlockBehaviour.PostProcess postProcess;
     @Shadow Predicate<BlockState> emissiveRendering;
     @Shadow boolean dynamicShape;
     @Shadow FeatureFlagSet requiredFeatures;
     @Shadow BlockBehaviour.OffsetFunction offsetFunction;
+    @Shadow float fallDistanceReduction;
 
+    @SuppressWarnings("deprecation")
     public BlockBehaviour.Properties copy() {
         BlockBehaviour.Properties properties = BlockBehaviour.Properties.of()
                 .mapColor(mapColor)
@@ -63,11 +69,14 @@ public abstract class MixinBlockBehaviorProperties implements IMixinBlockBehavio
                 .friction(friction)
                 .speedFactor(speedFactor)
                 .jumpFactor(jumpFactor)
+                .bounceRestitution(bounceRestitution)
+                .fallDistanceReduction(fallDistanceReduction)
                 .pushReaction(pushReaction)
                 .instrument(instrument)
                 .isValidSpawn(isValidSpawn)
                 .isRedstoneConductor(isRedstoneConductor)
                 .isSuffocating(isSuffocating)
+                .isViewBlocking(isViewBlocking)
                 .postProcess(postProcess)
                 .emissiveRendering(emissiveRendering);
 
@@ -79,15 +88,16 @@ public abstract class MixinBlockBehaviorProperties implements IMixinBlockBehavio
         if(ignitedByLava) { properties.ignitedByLava(); }
         if(liquid) { properties.liquid(); }
         if(forceSolidOn) { properties.forceSolidOn(); }
+        if(forceSolidOff) { properties.forceSolidOff(); }
         if(!spawnTerrainParticles) { properties.noTerrainParticles(); }
         if(replaceable) { properties.replaceable(); }
-        if(isViewBlocking != isSuffocating) { properties.isViewBlocking(isViewBlocking); }
         if(dynamicShape) { properties.dynamicShape(); }
 
         var mixinProperties = ((IMixinBlockBehaviorProperties) properties);
         mixinProperties.setFeatureFlagSet(requiredFeatures);
         mixinProperties.setOffsetFunction(offsetFunction);
         mixinProperties.overrideDrops(drops);
+        mixinProperties.overrideDescriptionId(descriptionId);
 
         return properties;
     }
@@ -105,5 +115,10 @@ public abstract class MixinBlockBehaviorProperties implements IMixinBlockBehavio
     @Override
     public void overrideDrops(DependantName<Block, Optional<ResourceKey<LootTable>>> drops) {
         this.drops = drops;
+    }
+
+    @Override
+    public void overrideDescriptionId(DependantName<Block, String> descriptionId) {
+        this.descriptionId = descriptionId;
     }
 }