+++ Ametrin/src/main/java/com/ametrinstudios/ametrin/util/mixin/IMixinBlockBehaviorProperties.java
@@ -17,4 +17,6 @@ public interface IMixinBlockBehaviorProperties {
     void setFeatureFlagSet(FeatureFlagSet featureFlags);
 
     void overrideDrops(DependantName<Block, Optional<ResourceKey<LootTable>>> drops);
+
+    void overrideDescriptionId(DependantName<Block, String> descriptionId);
 }