+++ Ametrin/src/main/java/com/ametrinstudios/ametrin/world/block/helper/BlockRegisterHelper.java
@@ -42,7 +42,7 @@ public final class BlockRegisterHelper {
         );
     }
     public static BlockBehaviour.Properties buttonProperties() {
-        return BlockBehaviour.Properties.of().noCollision().strength(0.5f).pushReaction(PushReaction.DESTROY);
+        return BlockBehaviour.Properties.of().noCollision().strength(0.5f).pushReaction(PushReaction.POPPED);
     }
 
     public static BlockBehaviour.Properties bushProperties() {
@@ -66,7 +66,7 @@ public final class BlockRegisterHelper {
     }
 
     public static BlockBehaviour.Properties pressurePlateProperties() {
-        return BlockBehaviour.Properties.of().forceSolidOn().noCollision().strength(0.5F).pushReaction(PushReaction.DESTROY);
+        return BlockBehaviour.Properties.of().forceSolidOn().noCollision().strength(0.5F).pushReaction(PushReaction.POPPED);
     }
 
     public static ToIntFunction<BlockState> litEmission(int lightLevel) {