+++ Ametrin/src/main/java/com/ametrinstudios/ametrin/world/block/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.ametrinstudios.ametrin.world.block;
+
+import org.jspecify.annotations.NullMarked;