+++ AmetrinTest/src/main/java/com/ametrinstudios/ametrin_test/AmetrinTestMod.java
@@ -1,17 +1,15 @@
 package com.ametrinstudios.ametrin_test;
 
-import com.ametrinstudios.ametrin.data.provider.CustomLootTableProvider;
 import com.ametrinstudios.ametrin_test.data.provider.*;
-import com.ametrinstudios.ametrin_test.data.provider.loot.TestBlockLootProvider;
-import com.ametrinstudios.ametrin_test.data.provider.loot.TestLootTableProvider;
 import com.ametrinstudios.ametrin_test.registry.TestBlocks;
 import com.ametrinstudios.ametrin_test.registry.TestItems;
 import com.ametrinstudios.ametrin_test.registry.TestPoiTypes;
 import com.mojang.logging.LogUtils;
+import net.minecraft.core.RegistrySetBuilder;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.Identifier;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
-import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.neoforged.neoforge.data.event.GatherDataEvent;
 import org.slf4j.Logger;
 
@@ -27,23 +25,17 @@ public final class AmetrinTestMod {
         TestItems.REGISTER.register(modBus);
         TestPoiTypes.REGISTER.register(modBus);
 
-        modBus.addListener(AmetrinTestMod::setup);
         modBus.addListener(AmetrinTestMod::gatherServerData);
         modBus.addListener(AmetrinTestMod::gatherClientData);
     }
 
-    private static void setup(final FMLCommonSetupEvent event) {
-//        VanillaCompat.addStrippable(TestBlocks.TEST_BLOCK.get(), Blocks.DIAMOND_BLOCK);
-//        VanillaCompat.addFlattenable(TestBlocks.TEST_BLOCK.get(), Blocks.DIAMOND_BLOCK);
-    }
-
     // Server data uses Client event because I don't need to run them separately
     public static void gatherServerData(GatherDataEvent.Client event) {
-        event.createProvider(CustomLootTableProvider.builder()
-                .addBlockProvider(TestBlockLootProvider::new)
-                .addChestProvider(TestLootTableProvider::new)::build);
+        event.createReloadableRegistryObjects(new RegistrySetBuilder()
+                .add(Registries.LOOT_TABLE, TestLootTableProvider.create())
+                .add(TestRecipeProvider.create())
+        );
 
-        event.createProvider(TestRecipeProvider.Runner::new);
         event.createProvider(TestBiomeTagsProvider::new);
         event.createBlockAndItemTags(TestBlockTagsProvider::new, TestItemTagsProvider::new);
     }