+++ AmetrinTest/src/main/java/com/ametrinstudios/ametrin_test/data/provider/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package com.ametrinstudios.ametrin_test.data.provider;
+
+import org.jspecify.annotations.NullMarked;