+++ changelog.md
@@ -1,3 +1,9 @@
+## 1.2.0
+- add `ExtendedLootSubProvider` (provides common `HolderGetter`s)
+- remove `CustomLootTableProvider` (use vanilla `LootTableProvider` directly)
+- remove `VanillaCompat.addFlowerPot`
+- disabled tree API (Vanilla tree generation should not be able to do the same)
+
 ## 1.1.0
 - data gen changes for vanilla compatibility
 - 26.2 port